M to M repo updates

+++ GirafRest/IRepositories/IRepositories M to M/IPictogramRelationRepository.cs
@@ -1,13 +1,10 @@
-using System;
+using GirafRest.Models;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq.Expressions;
-using GirafRest.Models; 
 
 namespace GirafRest.IRepositories
 {
     public interface IPictogramRelationRepository : IRepository<PictogramRelation>
     {
-        public ICollection<PictogramRelation> GetWithPictogram(int activityID);
+        public ICollection<PictogramRelation> GetWithPictogram(long activityID);
     }
 }
